Moved setup flow pages into their own files to reduce clutter.
